+++ src/bacnet/basic/service/h_awf.c
@@ -111,12 +111,21 @@ int handler_atomic_write_file_encode(
         pdu_len += len;
         return pdu_len;
     }
-    len = awf_decode_service_request(service_request, service_len, &data);
-    /* bad decoding - send an abort */
+    len = awf_decode_service_request(service_request, service_len, NULL);
     if (len < 0) {
+        /* bad decoding - send an abort */
         len = abort_encode_apdu(
             apdu, service_data->invoke_id, ABORT_REASON_OTHER, true);
         debug_print("AWF: Bad Encoding. Sending Abort!\n");
+        pdu_len += len;
+        return pdu_len;
+    }
+    len = awf_decode_service_request(service_request, service_len, &data);
+    if (len < 0) {
+        /* an input buffer capacity has been exceeded in this device */
+        len = abort_encode_apdu(
+            apdu, service_data->invoke_id, ABORT_REASON_BUFFER_OVERFLOW, true);
+        debug_fprintf(stderr, "AWF: Buffer Overflow. Sending Abort!\n");
         pdu_len += len;
         return pdu_len;
     }